pyproject.toml
[tool.poetry]
name = "tg"
version = "0.1.0"
description = ""
authors = ["ArtemKhel <ArtemKhel@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.4"
anytree = "2.8.0"

[tool.poetry.scripts]
tg = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
